Fixed metadata() in Opus. Needed to add public_path() since adding the slash to getImage(). Made favorite button partial Added favorite button to opus.show Refactored profile.show controller code. Now I'm thinking with laravel. Added getFilePath() to opus for the resizing function Added gitignore to art directory Navigator on galleryShow doesn't work

+++ app/Favorite.php
@@ -27,7 +27,7 @@ class Favorite extends Model
     
     public function add(User $user)
     {
-        if ($this->opus->user_id !== $user->id) {
+        if (!$user->isOwner($this->opus)) {
             $user->favorites()->attach($this->id);
             return true;
         }
@@ -36,15 +36,19 @@ class Favorite extends Model
 
     public function remove(User $user)
     {
-        $user->favorites()->detach($this->id);
+        if (!$user->isOwner($this->opus)) {
+            $user->favorites()->detach($this->id);
+            return true;
+        }
+        return false;
     }
 
     public static function has(User $user, Opus $opus)
     {
         $favorites = $user->favorites;
 
-        foreach($favorites as $favorite) {
-            if($favorite->opus_id === $opus->id) {
+        foreach ($favorites as $favorite) {
+            if ($favorite->opus_id === $opus->id) {
                 return true;
             }
         }
